Simplification of program
