+++ book_list.txt
@@ -1,0 +1,4 @@
+No.   | Book Name            | Author               | Price     
+-------------------------------------------------------------
+1     | Shirish ko Phool     | Bishnu Kumari Waiba  | RS 899       
+2     | One of Us is lying   | Karen M. McManu      | RS 1099
